+++ lib/flutter_stream_friends.dart
@@ -1,0 +1,4 @@
+library flutter_stream_friends;
+
+export 'src/flutter_stream_widget.dart';
+export 'src/flutter_stream_callbacks.dart';